+++ Johnson/docs/release-notes.md
@@ -2,6 +2,15 @@
 
 # Release notes
 
+## 1.0.1
+Released 2 February 2024.
+
+* Rebalanced Investment.
+* Second Quarter and Third Quarter now have their own art.
+* Mint effects can now be cancelled.
+* Mint A can now permanently upgrade temporarily upgraded cards.
+* Fixed duplicated card tooltips for all of the Quarter cards in the B upgrade.
+
 ## 1.0.0
 Released 1 February 2024.